Fix for webdriver_test_setup in Windows, and don't force close of the browsers if we're not running on a buildbot. Review URL: https://chromiumcodereview.appspot.com//9601023
git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@4984 260f80e4-7a28-3924-810f-c04153c831b5

+++ tools/testing/dart/test_runner.dart
@@ -905,7 +905,7 @@ class ProcessQueue {
    * and notify our progress indicator that we are done.
    */
   void _cleanupAndMarkDone() {
-    if (browserUsed != '') {
+    if (browserUsed != '' && _progress is BuildbotProgressIndicator) {
       killZombieBrowsers();
       if (_seleniumServer != null) {
         _seleniumServer.kill();
